Fix offline retry logic
refs: MBL-17022
affects: Student
release note: none
test plan: Start sync, turn Wi-Fi off, retry
